+++ tests/async_tests/test_decorator/test_collection_in_init.py
@@ -8,6 +8,7 @@ from tests.common import SimpleDTO, in_async_collection
 @pytest.mark.skip
 async def test_can_use_collection_in_init():
     async with in_async_collection(SimpleDTO) as collection:
+
         @mongorepo.use_collection(collection)
         @mongorepo.async_repository()
         class Repository: